content/media/test/test_wav_seek1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: seek test 1</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var startPassed = false;
var endPassed = false;
var seekFlagStart = false;
var seekFlagEnd = false;
var readonly = true;
var completed = false;

function startTest() {
  if (completed)
    return false;
  var v = document.getElementById('v');
  ok(!v.seeking, "seeking should default to false");
  try {
    v.seeking = 1;
    readonly = false;
  }
  catch(e) {
    readonly = true;
  }
  ok(readonly, "seeking should be readonly");

  v.play();
  v.currentTime=0.5;
  seekFlagStart = v.seeking;
  return false;
}

function seekStarted() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  v.pause();
  startPassed = true;
  return false;
}

function seekEnded() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  var t = v.currentTime;
  ok(t >= 0.4 && t <= 0.6, "Video currentTime should be around 0.5: " + t);
  v.play();
  endPassed = true;
  seekFlagEnd = v.seeking;
  return false;
}

function playbackEnded() {
  if (completed)
    return false;

  completed = true;
  ok(startPassed, "seeking event");
  ok(endPassed, "seeked event");
  ok(seekFlagStart, "seeking flag on start should be true");
  ok(!seekFlagEnd, "seeking flag on end should be false");
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<audio id='v'
       onended='return playbackEnded();'
       onloadedmetadata='return startTest();'
       onseeking='return seekStarted();'
       onseeked='return seekEnded();'>
  <source type='audio/x-wav' src='r11025_s16_c1.wav'>
</audio>
</body>
</html>
